move GooseRawRequest, GooseRequest into metrics.rs
